src/requirements.txt
python-dotenv
streamlit
google-generativeai
sqlite3
os
